requirements.txt
# Core framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.2.0

# Google APIs
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-auth==2.36.0

# Firebase for authentication
firebase-admin==6.6.0
google-cloud-firestore==2.21.0

# AI/LLM providers
anthropic==0.40.0
openai==1.58.1

# Scheduling
apscheduler==3.10.4

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
aiosqlite==0.20.0

# Authentication
pyjwt==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0

# Additional utilities
httpx==0.28.1
python-multipart==0.0.19
aiofiles==24.1.0
jinja2==3.1.5